+++ controller/OrderController.js
@@ -1,6 +1,7 @@
-//const conn = require('../mariadb'); // DB 모듈
 const mariadb = require('mysql2/promise');
 const {StatusCodes} = require('http-status-codes');  // status code 모듈
+const jwt = require('jsonwebtoken');
+const ensure_authorization = require('../auth');
 
 const order = async (req, res) => {
     const conn = await mariadb.createConnection({
@@ -10,38 +11,49 @@ const order = async (req, res) => {
         database : 'bookshop',
         dateStrings : true
     });   
-    
-    const {items, delivery, total_quantity, total_price, user_id, first_book_title} = req.body;
 
-    // delivery 테이블 삽입
-    let delivery_sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);`;
-    let delivery_values = [delivery.address, delivery.receiver, delivery.contact];
-    let [delivery_results] = await conn.execute(delivery_sql, delivery_values);
-    let delivery_id = delivery_results.insertId;
+    let authorization = ensure_authorization(req, res);
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message' : '로그인 세션이 만료되었습니다. 다시 로그인 하세요.'
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) { 
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message' : '잘못된 토큰입니다.'
+        });
+    } else {        
+        const {items, delivery, total_quantity, total_price, first_book_title} = req.body;
 
-    // orders 테이블 삽입
-    let orders_sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
-            VALUES (?, ?, ?, ?, ?);`;
-    let orders_values = [first_book_title, total_quantity, total_price, user_id, delivery_id]
-    let [orders_results] = await conn.execute(orders_sql, orders_values);
-    let order_id = orders_results.insertId;
-    
-    // items를 가지고 장바구니에서 book_id, quantity를 조회
-    let cart_items_sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?);`;
-    // let order_items = await conn.query(cart_items_sql, [items]);
-    let [order_items, fields] = await conn.query(cart_items_sql, [items]);
+        // delivery 테이블 삽입
+        let delivery_sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);`;
+        let delivery_values = [delivery.address, delivery.receiver, delivery.contact];
+        let [delivery_results] = await conn.execute(delivery_sql, delivery_values);
+        let delivery_id = delivery_results.insertId;
 
-    // orderedBook 테이블 삽입
-    let ordered_book_sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
-    // items는 배열로 받아오기 때문에 요소들을 하나씩 꺼냄
-    let ordered_book_values = order_items.map(item => [order_id, item.book_id, item.quantity]);
-    // 오류가 발생할 수 있으므로 query로 계속 진행
-    let ordered_book_results = await conn.query(ordered_book_sql, [ordered_book_values]); 
-    
-    // 주문 완료 후 cartItem 삭제
-    let result = await deleteCartItems(conn, items);
+        // orders 테이블 삽입
+        let orders_sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
+                VALUES (?, ?, ?, ?, ?);`;
+        let orders_values = [first_book_title, total_quantity, total_price, authorization.id, delivery_id]
+        let [orders_results] = await conn.execute(orders_sql, orders_values);
+        let order_id = orders_results.insertId;
+        
+        // items를 가지고 장바구니에서 book_id, quantity를 조회
+        let cart_items_sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?);`;
+        // let order_items = await conn.query(cart_items_sql, [items]);
+        let [order_items, fields] = await conn.query(cart_items_sql, [items]);
 
-    return res.status(StatusCodes.OK).json(result);
+        // orderedBook 테이블 삽입
+        let ordered_book_sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
+        // items는 배열로 받아오기 때문에 요소들을 하나씩 꺼냄
+        let ordered_book_values = order_items.map(item => [order_id, item.book_id, item.quantity]);
+        // 오류가 발생할 수 있으므로 query로 계속 진행
+        let ordered_book_results = await conn.query(ordered_book_sql, [ordered_book_values]); 
+        
+        // 주문 완료 후 cartItem 삭제
+        let result = await deleteCartItems(conn, items);
+
+        return res.status(StatusCodes.OK).json(result);
+    }
 };
 
 const deleteCartItems = async (conn, items) => {
@@ -60,17 +72,27 @@ const getOrders = async (req, res) => {
         database : 'bookshop',
         dateStrings : true
     });  
-
-    let sql = `SELECT orders.id, book_title, total_quantity, total_price, created_at,
-                address, receiver, contact
-                FROM orders LEFT JOIN delivery
-                ON orders.delivery_id = delivery.id;`;
-    let [rows, fields] = await conn.query(sql);
-    return res.status(StatusCodes.OK).json(rows);
+    let authorization = ensure_authorization(req, res);
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message' : '로그인 세션이 만료되었습니다. 다시 로그인 하세요.'
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) { 
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message' : '잘못된 토큰입니다.'
+        });
+    } else {   
+        let sql = `SELECT orders.id, book_title, total_quantity, total_price, created_at,
+                    address, receiver, contact
+                    FROM orders LEFT JOIN delivery
+                    ON orders.delivery_id = delivery.id;`;
+        let [rows, fields] = await conn.query(sql);
+        return res.status(StatusCodes.OK).json(rows);
+    }
 };
 
 const getOrderDetail =  async (req, res) => {
-    const {id} = req.params;
+    const order_id = req.params.id;
 
     const conn = await mariadb.createConnection({
         host : '127.0.0.1',
@@ -79,13 +101,23 @@ const getOrderDetail =  async (req, res) => {
         database : 'bookshop',
         dateStrings : true
     });   
-
-    let sql = `SELECT book_id, title, author, price, quantity
-                FROM orderedbook LEFT JOIN books
-                ON orderedbook.book_id = books.id
-                WHERE order_id = ?;`;
-    let [rows, fields] = await conn.query(sql, id);
-    return res.status(StatusCodes.OK).json(rows);
+    let authorization = ensure_authorization(req, res);
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message' : '로그인 세션이 만료되었습니다. 다시 로그인 하세요.'
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) { 
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            'message' : '잘못된 토큰입니다.'
+        });
+    } else {   
+        let sql = `SELECT book_id, title, author, price, quantity
+                    FROM orderedbook LEFT JOIN books
+                    ON orderedbook.book_id = books.id
+                    WHERE order_id = ?;`;
+        let [rows, fields] = await conn.query(sql, [order_id]);
+        return res.status(StatusCodes.OK).json(rows);
+    }
 };
 
 module.exports = {
